+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Enable anchor outputs on ln channels
+
 ## [1.5.1] - 2023-11-16
 
 - Add on-boarding wizard